requirements.txt
splinter[selenium]
bs4
python-telegram-bot[job-queue]